backend/requirements.txt
fastapi==0.128.1
uvicorn[standard]==0.40.0
sqlalchemy[asyncio]==2.0.38
asyncpg==0.30.0
pydantic-settings==2.12.0
redis==7.1.0
httpx==0.28.1
tenacity==9.0.0
python-multipart==0.0.22
python-dotenv==1.2.1
# Auth
PyJWT==2.11.0
bcrypt==4.2.1
# Testing & Security
pytest==8.3.4
pytest-asyncio==0.25.3
greenlet==3.1.1
bandit==1.8.3
pip-audit==2.8.0
# Payments & Tasks
stripe==11.5.0
celery==5.4.0

alembic==1.14.1
email-validator==2.2.0
